PIZ3 - Add Desing Patern (#4)
* refactor/JV: PIZ3 - Adding the factory method desing pattern

* feat/JV: PIZ3 - Created factory method desing pattern
