+++ app/factories/managers.py
@@ -1,5 +1,9 @@
-from app.repositories.managers import BaseManager, BeverageManager, IndexManager, IngredientManager, OrderManager, SizeManager
-
+from app.repositories.base_manager import BaseManager
+from app.repositories.beverage_manager import BeverageManager
+from app.repositories.index_manager import IndexManager
+from app.repositories.ingredient_manager import IngredientManager
+from app.repositories.order_manager import OrderManager
+from app.repositories.size_manager import SizeManager
 class ManagerFactory:
     _managers = {
         'base': BaseManager,
